+++ surprise/prediction_algorithms/algo_base.py
@@ -6,6 +6,9 @@ inherit.
 
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+import warnings
+
+from six import get_unbound_function as guf
 
 from .. import similarities as sims
 from .predictions import PredictionImpossible
@@ -14,7 +17,7 @@ from .optimize_baselines import baseline_als
 from .optimize_baselines import baseline_sgd
 
 
-class AlgoBase:
+class AlgoBase(object):
     """Abstract class where is defined the basic behavior of a prediction
     algorithm.
 
@@ -31,8 +34,24 @@ class AlgoBase:
         self.sim_options = kwargs.get('sim_options', {})
         if 'user_based' not in self.sim_options:
             self.sim_options['user_based'] = True
+        self.skip_train = False
+
+        if (guf(self.__class__.fit) is guf(AlgoBase.fit) and
+           guf(self.__class__.train) is not guf(AlgoBase.train)):
+            warnings.warn('It looks like this algorithm (' +
+                          str(self.__class__) +
+                          ') implements train() '
+                          'instead of fit(): train() is deprecated, '
+                          'please use fit() instead.', UserWarning)
 
     def train(self, trainset):
+        '''Deprecated method: use :meth:`fit() <AlgoBase.fit>`
+        instead.'''
+
+        self.skip_train = True
+        self.fit(trainset)
+
+    def fit(self, trainset):
         """Train an algorithm on a given training set.
 
         This method is called by every derived class as the first basic step
@@ -44,6 +63,24 @@ class AlgoBase:
                 set, as returned by the :meth:`folds
                 <surprise.dataset.Dataset.folds>` method.
         """
+
+        # Check if train method is overridden: this means the object is an old
+        # style algo (new algo only have fit() so self.__class__.train will be
+        # AlgoBase.train). If true, there are 2 possible cases:
+        # - algo.fit() was called. In this case algo.train() was skipped which
+        #   is bad. We call it and skip this part next time we enter fit().
+        #   Then return immediatly because fit() has already been called by
+        #   AlgoBase.train() (which has been called by algo.train()).
+        # - algo.train() was called, which is the old way. In that case,
+        #   the skip flag will ignore this.
+        # This is fairly ugly and hacky but I did not find anything better so
+        # far, in order to maintain backward compatibility... See
+        # tests/test_train2fit.py for supported cases.
+        if (guf(self.__class__.train) is not guf(AlgoBase.train) and
+                not self.skip_train):
+            self.train(trainset)
+            return
+        self.skip_train = False
 
         self.trainset = trainset